+++ docs/docs/search/functions_1.js
@@ -1,0 +1,7 @@
+var searchData=
+[
+  ['defaultmongodbfactory',['DefaultMongoDbFactory',['../class_pip_services3_1_1_mongo_db_1_1_build_1_1_default_mongo_db_factory.html#aae012159819de2395b42902f57d7e37f',1,'PipServices3::MongoDb::Build::DefaultMongoDbFactory']]],
+  ['deletebyfilterasync',['DeleteByFilterAsync',['../class_pip_services3_1_1_mongo_db_1_1_persistence_1_1_mongo_db_persistence.html#a1fa8210d4f007d68ac3829ea9bc881c3',1,'PipServices3::MongoDb::Persistence::MongoDbPersistence']]],
+  ['deletebyidasync',['DeleteByIdAsync',['../class_pip_services3_1_1_mongo_db_1_1_persistence_1_1_identifiable_mongo_db_persistence.html#a16034934916fb5fa778652217e500df7',1,'PipServices3.MongoDb.Persistence.IdentifiableMongoDbPersistence.DeleteByIdAsync()'],['../class_pip_services3_1_1_mongo_db_1_1_persistence_1_1_partition_mongo_db_persistence.html#a76877a2d20b8765f27aeb71bd256dbfa',1,'PipServices3.MongoDb.Persistence.PartitionMongoDbPersistence.DeleteByIdAsync()']]],
+  ['deletebyidsasync',['DeleteByIdsAsync',['../class_pip_services3_1_1_mongo_db_1_1_persistence_1_1_identifiable_mongo_db_persistence.html#a671016a9bef47aefd62c00cc0273a667',1,'PipServices3::MongoDb::Persistence::IdentifiableMongoDbPersistence']]]
+];